move UserRoleInfo, SeriesRoleInfo to reusable RoleInfo component
